feat: extract @mizchi/playwright-faults (network / lifecycle / runtime)
Tier 2 Layer-1 extraction following #43 (server-faults), #50
(playwright-v8-coverage). Pulls all fault-injection primitives out of
chaosbringer into a single package usable by any Playwright project.

What moves (3 source + 3 test + types):
- runtime-faults.ts (256) -> packages/playwright-faults/src/
- runtime-faults.test.ts (282) -> packages/playwright-faults/src/
- lifecycle-faults.ts (252) -> packages/playwright-faults/src/
- lifecycle-faults.test.ts (252) -> packages/playwright-faults/src/
- faults.ts (225) -> packages/playwright-faults/src/
- faults.test.ts (127) -> packages/playwright-faults/src/

What moves (types from chaosbringer/src/types.ts):
- Fault, FaultRule, FaultInjectionStats, UrlMatcher (network layer)
- LifecycleStage, StorageScope, LifecycleAction, LifecycleFault,
  LifecycleFaultStats (lifecycle layer)
- RuntimeAction, RuntimeFault, RuntimeFaultStats (runtime layer)
- New: minimal Rng interface (just `next(): number`) so the package
  is seed-agnostic; chaosbringer's createRng remains structurally
  compatible

chaosbringer/src/types.ts re-exports the fault types from
@mizchi/playwright-faults so CrawlerOptions and CrawlReport
references continue to resolve. chaosbringer/src/{runtime,lifecycle,}faults.ts
become thin re-export shims — pure refactor, zero behavior change.

Workspace dep added; tsc topological build via prepare scripts (the
recipe established in #50).

Verified locally:
- pnpm install (fresh, no node_modules) -> 4 packages prepare cleanly
- pnpm test (recursive) -> 487/487 (chaosbringer 391 + playwright-faults
  66 + playwright-v8-coverage 17 + server-faults 13)
- NUL gate clean for all packages

Publish ordering: not yet on npm. First publish needs manual mizchi
action (OIDC bootstrap). chaosbringer's next release after this PR
must wait for @mizchi/playwright-faults to be on npm — same caveat
as #50.
